alrj-solution: removed unneeded import & unneeded test case, updated tests and README.md, comments updated to more formal style
