+++ Java/src/main/java/com/gildedrose/GildedRose.java
@@ -1,6 +1,5 @@
 package com.gildedrose;
 
-import java.lang.reflect.Array;
 import java.util.Arrays;
 
 class GildedRose {
