+++ Java/src/main/java/com/gildedrose/NormallyProcessedItem.java
@@ -9,10 +9,17 @@ public class NormallyProcessedItem extends Item {
         super(name, sellIn, quality);
     }
 
+    /**
+     * The quality of the item decreases by 1 for each decrease in the sellIn value
+     * The minimum quality is 0
+     */
     public void updateQuality() {
         decreaseQuality(1);
     }
 
+    /**
+     * The sellIn value decreases by 1 each day
+     */
     public void updateSellIn() {
         sellIn -= 1;
     }
